requirements.txt
flask
flask-sqlalchemy
nltk
scikit-learn
tensorflow==2.16.1
keras==3.3.3
Pillow